fix: move Checkpoint to signer module out of simulation
