+++ .github/workflows/ci.yml
@@ -17,7 +17,6 @@ jobs:
         os:
           - macos-latest
           - ubuntu-latest
-          - wasm32-unknown-unknown
           - windows-latest
         channel:
           - stable
